Refine defaultModuleName logic. Move to generate.go.
