+++ gazelle/generate.go
@@ -56,11 +56,34 @@ func genRulesFromSrcFiles(sc *swiftcfg.SwiftConfig, args language.GenerateArgs) 
 	sort.Strings(srcs)
 
 	// Generate the rules and imports
-	result.Gen = swift.RulesFromSrcs(args, srcs)
+	defaultModuleName := defaultModuleName(args)
+	result.Gen = swift.RulesFromSrcs(args, srcs, defaultModuleName)
 	result.Imports = swift.Imports(result.Gen)
 	result.Empty = generateEmpty(args, srcs)
 
 	return result
+}
+
+func defaultModuleName(args language.GenerateArgs) string {
+	// Order of names to use
+	// 1. Value specified via directive.
+	// 2. Directory name.
+	// 3. Repository name.
+	// 4. "DefaultModule"
+
+	// Check for a value configured via directive
+	sc := swiftcfg.GetSwiftConfig(args.Config)
+	if defaultModuleName, ok := sc.DefaultModuleNames[args.Rel]; ok {
+		return defaultModuleName
+	}
+	defaultModuleName := filepath.Base(args.Config.WorkDir)
+	if defaultModuleName == "." || defaultModuleName == "" {
+		defaultModuleName = args.Config.RepoName
+	}
+	if defaultModuleName == "" {
+		defaultModuleName = "DefaultModule"
+	}
+	return defaultModuleName
 }
 
 // Look for any rules in the existing BUILD file that do not reference one of the source files. If
